pyproject.toml
[tool.poetry]
name = "MushiCord"
version = "0.1.0"
description = "A super EPIC Music Discord bot made by Yoshiboi18303#4045 that can help you play many songs from YouTube with just a URL!"
authors = ["Yoshiboi18303 <yoshiboi18303.t@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
youtube_dl = "^2021.6.6"
PyNaCl = "^1.4.0"
Flask = "^2.0.1"
pymongo = {extras = ["srv"], version = "^3.12.0"}
"statcord.py" = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
